cloud-proxy/requirements.txt
# Cloud Run Proxy Dependencies
Flask==2.3.3
flask-cors==4.0.0
google-auth==2.23.4
requests==2.31.0
numpy==1.24.3
redis==5.0.1
gunicorn==21.2.0